:recycle: (hooks) rename and improve hooks
