+++ source/layout/SettingsModal.tsx
@@ -4,8 +4,8 @@ import tw from 'twin.macro';
 import { ReactComponent as EditIcon } from '@/assets/icons/edit.svg';
 import { Button } from '@/components/Button';
 import { Input } from '@/components/Input';
+import { ModalContext } from '@/context/modal';
 import { t } from '@/helpers/translate';
-import { ModalContext } from '@/hooks/context/modal';
 import { useUpdatePlayer } from '@/services/player/mutations';
 import { useSession } from '@/services/player/queries';
