+++ source/pages/Room/Playing/PlayerKilledButton.tsx
@@ -3,8 +3,8 @@ import tw from 'twin.macro';
 
 import Knife from '@/assets/images/knife.png';
 import { Button } from '@/components/Button';
+import { ModalContext } from '@/context/modal';
 import { t } from '@/helpers/translate';
-import { ModalContext } from '@/hooks/context/modal';
 
 import { PlayerKilledModal } from './PlayerKilledModal';
